ingestion_service/requirements.txt
pillow
fastapi
uvicorn
requests
kafka-python
easyocr
